// src/servers/ecommerce/types/Cart.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Cart = {
  /**
   * @description ID of the cart
   * @type string
   */
  cartId: string
  /**
   * @description ID of the customer who owns the cart
   * @type string
   */
  customerId: string
  /**
   * @description List of products in the cart
   * @type array
   */
  products: {
    /**
     * @description ID of the product
     * @type string | undefined
     */
    productId?: string
    /**
     * @description Name of the product
     * @type string | undefined
     */
    name?: string
    /**
     * @description Price of the product
     * @type number | undefined
     */
    price?: number
    /**
     * @description Photos of the product
     * @type array | undefined
     */
    photos?: string[]
    /**
     * @description Quantity of the product in the cart
     * @type integer | undefined
     */
    quantity?: number
  }[]
  /**
   * @description Total price of the cart
   * @type number
   */
  shipPrice: number
}