// src/servers/ecommerce/types/Product.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Product = {
  /**
   * @description ID of the product
   * @type string
   */
  productId: string
  /**
   * @description Name of the product
   * @type string
   */
  name: string
  /**
   * @description Price of the product
   * @type number
   */
  price: number
  /**
   * @description Description of the product
   * @type string | undefined
   */
  description?: string
  /**
   * @description Photos of the product
   * @type array | undefined
   */
  photos?: string[]
  /**
   * @description Category of the product
   * @type object | undefined
   */
  category?: {
    /**
     * @description ID of the category
     * @type string | undefined
     */
    categoryId?: string
    /**
     * @description Name of the category
     * @type string | undefined
     */
    name?: string
  }
  /**
   * @description Subcategory of the product
   * @type object | undefined
   */
  subcategory?: {
    /**
     * @description ID of the subcategory
     * @type string | undefined
     */
    subCategoryId?: string
    /**
     * @description Name of the subcategory
     * @type string | undefined
     */
    name?: string
  }
}